+++ models/sipnet/tests/testthat/test-model2netcdf.SIPNET.R
@@ -44,7 +44,7 @@ make_base_sipnet <- function(n = 4L) {
     year = 2002,
     day = rep(c(1, 2), each = n / 2, length.out = n),
     time = rep(c(6, 18), length.out = n),
-    plantWoodC = 5000, plantLeafC = 200, woodCreation = 0.5,
+    plantWoodC = 5000, plantLeafC = 200,
     soil = 10000, microbeC = 8, coarseRootC = 1200, fineRootC = 800,
     litter = 400, soilWater = 14, soilWetnessFrac = 0.85, snow = 0,
     npp = 0.05, nee = 0.10, cumNEE = cumsum(rep(0.1, n)),
@@ -226,4 +226,54 @@ test_that("fluxes are converted from gC/m2/timestep to kg/m2/sec", {
 
   expect_equal(pec$GPP, sip$gpp / 1000 / ts)
   expect_equal(pec$Transp, sip$fluxestranspiration * 10 / ts, tolerance = 1e-6)
+
+  sip2 <- make_v2_sipnet()
+  out2 <- out_dir <- setup_sipnet_test(sip2)$outdir
+  pec2 <- PEcAn.utils::read.output(
+    ncfiles = file.path(out2, "2002.nc"),
+    variables = c("GPP", "GWBI", "Transp"),
+    dataframe = TRUE,
+    verbose = FALSE,
+    print_summary = FALSE
+  )
+  expect_equal(pec2$GWBI, sip2$woodCreation / 1000 / ts, tolerance = 1e-6)
 })
+
+
+test_that("sipnet2datetime - standard vectorised input", {
+  years <- c(2023, 2023)
+  doys <- c(1, 32)
+  hours <- c(0, 10.5)
+
+  datetimes <- sipnet2datetime(years, doys, hours)
+
+  expect_equal(length(datetimes), 2)
+  expect_equal(datetimes[1], as.POSIXct("2023-01-01 00:00:00", tz = "UTC"))
+  expect_equal(datetimes[2], as.POSIXct("2023-02-01 10:30:00", tz = "UTC"))
+  }
+)
+
+test_that("sipnet2datetime - leap years",{
+
+  expect_equal(
+    format(sipnet2datetime(2024, 60, 0), "%Y-%m-%d"), "2024-02-29")
+
+  expect_equal(
+    format(sipnet2datetime(2023, 60, 0), "%Y-%m-%d"), "2023-03-01")
+  }
+)
+
+test_that("sipnet2datetime - decimal accuracy", {
+  expect_equal(format(sipnet2datetime(2023, 1, 13.75), "%H:%M:%S"),
+               "13:45:00")
+
+  expect_equal(format(sipnet2datetime(2023, 1, 23.9999), "%Y-%m-%d %H:%M:%S"),
+               "2023-01-01 23:59:59")
+
+  }
+)
+
+test_that("sipnet2datetime - UTC timezone", {
+  expect_equal(attr(sipnet2datetime(2023, 1, 1), "tzone"), "UTC")
+  }
+)